VACMS-8499 update token sb pages
