+++ docroot/design-system/.storybook/cssVariables.json
@@ -53,6 +53,14 @@
     "--spacing-l": "1.5rem",
     "--spacing-ll": "2rem",
     "--spacing-xl": "3rem",
+    "--va-display-xl": "48px",
+    "--va-display-l": "40px",
+    "--va-display-m": "32px",
+    "--va-display-s": "24px",
+    "--va-subsection-s": "18px",
+    "--va-subsection-xs": "15px",
+    "--va-text": "18px",
+    "--va-helptext": "13px",
     "--color-absolutezero": "var(--va-blue-med)",
     "--color-absolutezero-hover": "var(--va-blue-dark)",
     "--color-absolutezero-active": "var(--va-blue-darker)",
