+++ docroot/design-system/components/tokens/_variables.css
@@ -56,6 +56,15 @@
   --spacing-l: 1.5rem;
   --spacing-ll: 2rem;
   --spacing-xl: 3rem;
+  /* font sizes */
+  --va-display-xl: 48px;
+  --va-display-l: 40px;
+  --va-display-m: 32px;
+  --va-display-s: 24px;
+  --va-subsection-s: 18px;
+  --va-subsection-xs: 15px;
+  --va-text: 18px;
+  --va-helptext: 13px;
   /*
    * Variables originally defined in core/themes/claro (or other modules), overridden here.
    */
