+++ docroot/design-system/components/tokens/colors.stories.js
@@ -1,0 +1,35 @@
+import colors from "./sb-colors.twig";
+
+// These are our design system tokens, generated by `yarn css`
+import customProperties from '../../.storybook/cssVariables.json'
+const vars = customProperties["custom-properties"];
+
+// Filter the CSS properties we want for a given Story
+const colorVars = Object.keys(vars)
+  .filter(key => key.includes('--va-'))
+  .reduce((obj, key) => {
+    if (vars[key].toLowerCase().match('#')) {
+      obj[key] = vars[key];
+    }
+    return obj;
+  }, {});
+
+export default { title: 'Design System Tokens' }
+
+export const Colors = {
+  args: {
+    vars: colorVars
+  },
+  parameters: {
+    options: { showPanel: false }
+  },
+  argTypes: {
+    vars: {
+      table: { disable: true }
+    }
+  }
+}
+
+Colors.parameters.render = async args => {
+  return await colors({...Colors.args, ...args});
+}
