--- docroot/design-system/components/tokens/colors.twig
@@ -1,5 +1,0 @@
-<div class="color-swatches">
-  {% for name, value in vars %}
-    <div>{{ name }} - {{ value }}</div>
-  {% endfor %}
-</div>
