+++ docroot/design-system/components/tokens/sb-typography.twig
@@ -1,0 +1,10 @@
+<div class="typography">
+  {% for name, value in vars %}
+    <div>{{ name }} - {{ value }}</div>
+  {% endfor %}
+</div>
+
+
+{% for name, value in sizes %}
+  <div>{{ name }} - {{ value }}</div>
+{% endfor %}
