+++ docroot/design-system/components/tokens/spacing.stories.js
@@ -1,0 +1,33 @@
+import spacing from "./sb-spacing.twig";
+
+// These are our design system tokens, generated by `yarn css`
+import customProperties from '../../.storybook/cssVariables.json'
+const vars = customProperties["custom-properties"];
+
+// Filter the CSS properties we want for a given Story
+const spacingVars = Object.keys(vars)
+  .filter(key => key.includes('--spacing'))
+  .reduce((obj, key) => {
+    obj[key] = vars[key];
+    return obj;
+  }, {});
+
+export default { title: 'Design System Tokens' }
+
+export const Spacing = {
+  args: {
+    vars: spacingVars
+  },
+  parameters: {
+    options: { showPanel: false }
+  },
+  argTypes: {
+    vars: {
+      table: { disable: true }
+    }
+  }
+}
+
+Spacing.parameters.render = async args => {
+  return await spacing({...Spacing.args, ...args});
+}
