--- docroot/design-system/components/tokens/spacing.twig
@@ -1,5 +1,0 @@
-<div class="">
-  {% for name, value in vars %}
-    <div>{{ name }} - {{ value }}</div>
-  {% endfor %}
-</div>
