+++ docroot/design-system/components/tokens/typography.stories.js
@@ -1,0 +1,43 @@
+import typography from './sb-typography.twig';
+
+// These are our design system tokens, generated by `yarn css`
+import customProperties from '../../.storybook/cssVariables.json'
+const vars = customProperties["custom-properties"];
+
+const typeVars = Object.keys(vars)
+  .filter(key => key.includes('--font'))
+  .reduce((obj, key) => {
+    obj[key] = vars[key];
+    return obj;
+  }, {});
+
+// Filter the CSS properties we want for a given Story
+const sizeVars = Object.keys(vars)
+  .filter(key => key.includes('--va-'))
+  .reduce((obj, key) => {
+    if (vars[key].toLowerCase().match('px')) {
+      obj[key] = vars[key];
+    }
+    return obj;
+  }, {});
+
+export default { title: 'Design System Tokens' }
+
+export const Typography = {
+  args: {
+    vars: typeVars,
+    sizes: sizeVars
+  },
+  parameters: {
+    options: { showPanel: false }
+  },
+  argTypes: {
+    vars: {
+      table: { disable: true }
+    }
+  }
+}
+
+Typography.parameters.render = async args => {
+  return await typography({...Typography.args, ...args});
+}
