--- docroot/design-system/components/tokens/typography.twig
@@ -1,5 +1,0 @@
-<div class="">
-  {% for name, value in vars %}
-    <div>{{ name|raw }} - {{ value|raw }}</div>
-  {% endfor %}
-</div>
